+++ modules/circe/src/main/scala/io/bartholomews/fsclient/circe/FsClientCirceApi.scala
@@ -50,7 +50,7 @@ trait FsClientCirceApi extends SttpCirceApi {
       accessToken <- c.downField("access_token").as[AccessToken]
       tokenType <- c.downField("token_type").as[String]
       expiresIn <- c.downField("expires_in").as[Long]
-      refreshToken <- c.downField("refresh_token").as[Option[RefreshToken]]
+      refreshToken <- c.downField("refresh_token").as[RefreshToken]
       scope <- c.downField("scope").as[Scope]
     } yield AccessTokenSigner(
       generatedAt.getOrElse(System.currentTimeMillis()),