+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Removed
 ### Changed
-- `AccessTokenSigner` using `currentTimeMillis` instead of `nanoTime`
+- `AccessTokenSigner` using `currentTimeMillis` instead of `nanoTime`, and has `refreshToken` required
 - `baseRequest` taking an `userAgent` instead of `client`
 ### Fixed